Changed selected tile to derive from primary geometry
Selected tile was being derived from the dual topology,
(Which is redundant and shouldn't be generated!!!), which
meant that it had way more triangles than it should have done.

Note that the selected tile's vertex format is currently identical to
the Primary, which is defined by the initial IcoSphere generator
function to be Vertex3DColor.